pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["version"]
name = "msglc"
description = "msgpack with lazy/partial loading containers"
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
keywords = ["msgpack", "serialization", "lazy loading"]
authors = [{ name = "Theodore Chang", email = "tlcfem@gmail.com" }]
maintainers = [{ name = "Theodore Chang", email = "tlcfem@gmail.com" }]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
    "msgpack>=1",
]

[project.optional-dependencies]
test = [
    "pytest-cov",
    "black",
]

[project.urls]
"Homepage" = "https://github.com/TLCFEM/msglc"
"Bug Reports" = "https://github.com/TLCFEM/msglc/issuess"
"Source" = "https://github.com/TLCFEM/msglc"

[tool.black]
line-length = 120
fast = true
